presidio-image-redactor/requirements.txt
flask==1.1.2
pillow
pytesseract
presidio-analyzer